Text Augmentation additions
Add two new modules for modifying text in captions:

CapitalizeTags: Will randomly capitalize tags in the caption with a set probability. Can specify comma-separated list of capitalization modes - 'capslock' (ALL CAPS), 'title' (First Letter Of Every Word), 'first' (First word only), or 'random' (rAnDOm lETteRS) and it will iterate through each tag in the caption and apply a random mode to each. Also has an option to set the entire caption to lowercase first. Uses the same delimiter as defined for the existing "shuffle tags" option.

DropTags: Will randomly remove tags from the caption with a specified probability. Also currently uses the same existing "keep tags" and "delimiter" options. Has three different modes of dropping tags:
   - FULL - all-or-nothing, will drop everything after the 'keep tags' amount
   - RANDOM - will iterate through the caption and randomly drop individual tags
   - RANDOM WEIGHTED - same as random, but the probability to drop tags is reduced at the start of the caption and only reaches the full value at the end, making it more likely to preserve tags at the start

Also supports a "special tags" list, which can be either a delimiter-separated string in the input field or a file path to a txt or csv file with entries separated by new lines. Can act as a "whitelist" (the special tags will never be dropped, but all others past the "keep tags" amount can be) or "blacklist" (the special tags may be dropped but all others will always be kept). There's also an option to enable this list to be interpreted with regex matching, so that syntax like ".*s" will match any tags ending with the letter s. Includes an exception for the "/(" and "/)" syntax used in many booru/e6 tags, but there may be some other special characters used by regex "[]\.^$*+?{}|()" which could be interpreted in unintentional ways.