Change all_possible_input_aspects to attribute
Changed all_possible_instance_aspects to class attribute in AspectBucketing, allows the list to be accessed from other functions. For the "concept statistics" I'm working on in OneTrainer this will let it pull the values and create all possible bucket resolutions without hardcoding them, preventing issues if they are changed in the future. Should not change the behavior of the existing class.